Fix fresh MySQL installation issues
- Fix foreign key type mismatch in appointments table migration
  (company_id and creator_id should be unsignedInteger, not unsignedBigInteger)
- Add try-catch in routes/console.php to handle RecurringInvoice query
  during migrations when deleted_at column doesn't exist yet

+++ database/migrations/2025_11_15_190401_create_appointments_table.php
@@ -13,9 +13,9 @@ return new class extends Migration
     {
         Schema::create('appointments', function (Blueprint $table) {
             $table->id();
-            $table->unsignedBigInteger('customer_id');
-            $table->unsignedBigInteger('company_id');
-            $table->unsignedBigInteger('creator_id')->nullable();
+            $table->unsignedBigInteger('customer_id');  // Match customers.id ($table->id() = unsigned BIGINT)
+            $table->unsignedInteger('company_id');  // Match companies.id (increments = unsigned INT)
+            $table->unsignedInteger('creator_id')->nullable();  // Match users.id (increments = unsigned INT)
             
             // Unique hash for secure identification
             $table->string('unique_hash', 50)->nullable();
